finished construction of mPets and mBundles
member variables of the class MainWindow mPets and mBundles are now complete. added destructors to petBuilder and bundleBuilder to eliminate memory leaks. commented

+++ bird.cpp
@@ -18,11 +18,11 @@ QString Bird::GetName() const{
 QString Bird::GetType() const{
     return this->_type;
 }
-unsigned int Bird::GetWeight() const{
+double Bird::GetWeight() const{
     return this->_weight;
 }
  //getter for price
-float Bird::GetPrice() const{
+double Bird::GetPrice() const{
     return this->_price;
 }
 // getter for nocturnal
