+++ bird.h
@@ -17,9 +17,9 @@ public:
     virtual QString GetName() const;
     // getter for type
     virtual QString GetType() const;
-    virtual unsigned int GetWeight() const;
+    virtual double GetWeight() const;
      //getter for price
-    virtual float GetPrice() const;
+    virtual double GetPrice() const;
     // getter for nocturnal
     bool GetNocturnal() const;
 };
